backend/pyproject.toml
[project]
name = "uniknow-backend"
version = "0.1.0"
description = "Case management knowledge base SaaS system with GraphRag Q&A"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langgraph>=0.0.15",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.3",
    "langchain-community>=0.0.10",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "motor>=3.3.2",
    "redis>=5.0.1",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
